# specs/001-concierge-chat-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Aueshah AI Concierge API
  version: 2.0.0
  description: |
    Phase 2 API — adds WordPress-bridged authentication, persistence (Neon),
    rate limiting, Noor allocation workflow, appointments with notifications,
    and concierge admin endpoints.

    Phase 1 endpoints (`POST /chat`, `GET /health`) remain backward-compatible.
  contact:
    name: Aueshah Engineering
    email: engineering@aueshah.com
servers:
  - url: https://api.aueshah.com
    description: Production
  - url: http://localhost:8000
    description: Local dev
security:
  - BearerAuth: []
...
paths:
...
  /v1/auth/wp-login:
    post:
      tags: [auth]
      security: []
      summary: Exchange a WordPress JWT for our session JWT
      description: |
        Verifies the WP-issued JWT (RS256, via cached JWKS), upserts the user
        into Neon, and returns our session access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wp_token]
              properties:
                wp_token: { type: string, description: JWT issued by the WP JWT plugin }
                visitor_id: { type: string, format: uuid, description: Anonymous visitor cookie to merge }
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, example: Bearer }
                  expires_in: { type: integer, example: 86400 }
                  user: { $ref: '#/components/schemas/UserPublic' }
        '401': { $ref: '#/components/responses/AuthFailure' }
  /v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPublic' }
        '401': { $ref: '#/components/responses/AuthFailure' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
  schemas:
...
    UserPublic:
      type: object
      properties:
        id: { type: string, format: uuid }
        wp_user_id: { type: integer }
        email: { type: string, format: email }
        display_name: { type: string }
        age_range: { type: string }
        skin_tone: { type: string }
        style_preference: { type: string }
        preferred_collection: { type: string }
        favorite_metals: { type: array, items: { type: string } }
        favorite_styles: { type: array, items: { type: string } }
        role: { type: string, enum: [client, concierge, admin] }

    # Appointment
...
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: integer }
        retry_after: { type: integer, description: Seconds (429/408 only) }
  responses:
...
    AuthFailure:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...